total refactor of functions
