--- deid/identifiers/README.md
@@ -1,3 +1,0 @@
-# Identifiers
-
-This module is intended for general functions related to identifiers, regardless of modality.
