--- deid/identifiers/__init__.py
@@ -1,7 +1,0 @@
-from .utils import (
-    save_identifiers,
-    load_identifiers,
-    get_timestamp
-)
-
-from .tasks import clean_identifiers
