+++ docs/_docs/getting-started/dicom-put.md
@@ -1,7 +1,7 @@
 ---
 title: 5. Put Identifiers Back
 category: Getting Started
-order: 6
+order: 7
 ---
