--- docs/_docs/tools/index.md
@@ -1,16 +1,0 @@
----
-title: Overview
-category: Tools
-permalink: /tools/index.html
-order: 1
----
-
-The following are helpful functions and tools that might be of interest.
-
- - [Tags]({{ site.baseurl }}/tools/tags): A few helpful functions for searching and filtering tags.
-Penrose is software that enables people to <b>create beautiful diagrams 
-just by typing mathematical notation in plain text.</b> The goal is to make it 
-easy for non-experts to create and explore high-quality diagrams, providing 
-deeper insight into challenging technical concepts. We aim to <b>democratize 
-the process of creating visual intuition</b>.
-
