+++ docs/_docs/user-docs/index.md
@@ -1,0 +1,27 @@
+---
+title: Documentation
+category: User Documentation
+permalink: /user-docs/index.html
+order: 1
+---
+
+Along with the [getting started]({{ site.baseurl }}/getting-started/) guides,
+these pages will help you to use the deid software. 
+
+## Recipes
+
+ - [Filters]({{ site.baseurl }}/user-docs/recipe-filters/): How to write sections to filter and flag images.
+ - [Headers]({{ site.baseurl }}/user-docs/recipe-headers/): How to write header actions to update image headers.
+ - [Labels]({{ site.baseurl }}/user-docs/recipe-labels/): can be used to add metadata to your recipes.
+
+## Client
+
+ - [Client]({{ site.baseurl }}/user-docs/client): A command line client for basic cleaning.
+
+
+## Tools
+
+ - [Tags]({{ site.baseurl }}/user-docs/tags): A few helpful functions for searching and filtering tags.
+
+
+
